fix: update and init with resource

+++ internal/controller/k8s/k8scluster/cluster.go
@@ -232,6 +232,9 @@ func (c *externalCluster) Delete(ctx context.Context, mg resource.Managed) error
 	if cr.Status.AtProvider.State == compute.DESTROYING || cr.Status.AtProvider.State == k8s.TERMINATED {
 		return nil
 	}
+	if cr.Status.AtProvider.State != compute.ACTIVE {
+		return fmt.Errorf("resource needs to be in ACTIVE state to delete it, current state: %v", cr.Status.AtProvider.State)
+	}
 
 	apiResponse, err := c.service.DeleteK8sCluster(ctx, cr.Status.AtProvider.ClusterID)
 	if err != nil {
